Move sleep at eoc state to sleepTimer
