+++ playback/src/main/kotlin/voice/playback/session/SleepTimer.kt
@@ -5,6 +5,7 @@ import kotlin.time.Duration
 
 interface SleepTimer {
   val leftSleepTimeFlow: Flow<Duration>
+  var sleepAtEoc: Boolean
   fun sleepTimerActive(): Boolean
   fun setActive(enable: Boolean)
   fun setEocActive(enable: Boolean)
