+++ sleepTimer/src/main/kotlin/voice/sleepTimer/SleepTimer.kt
@@ -48,9 +48,20 @@ class SleepTimer
     }
   override val leftSleepTimeFlow: StateFlow<Duration> get() = _leftSleepTime
 
-  override fun sleepTimerActive(): Boolean = (sleepJob?.isActive == true && leftSleepTime > Duration.ZERO) || playStateManager.sleepAtEoc
+  override fun sleepTimerActive(): Boolean = (sleepJob?.isActive == true && leftSleepTime > Duration.ZERO) || sleepAtEoc
 
   private var sleepJob: Job? = null
+
+  private val _sleepAtEoc = MutableStateFlow(false)
+
+  val sleepAtEocFlow: StateFlow<Boolean>
+    get() = _sleepAtEoc
+
+  override var sleepAtEoc: Boolean
+    set(value) {
+      _sleepAtEoc.value = value
+    }
+    get() = _sleepAtEoc.value
 
   override fun setActive(enable: Boolean) {
     Logger.i("enable=$enable")
@@ -60,9 +71,10 @@ class SleepTimer
       cancel()
     }
   }
+
   override fun setEocActive(enable: Boolean) {
     if (enable) {
-      playStateManager.sleepAtEoc = true
+      sleepAtEoc = true
     } else {
       cancel()
     }
@@ -133,6 +145,6 @@ class SleepTimer
     sleepJob?.cancel()
     leftSleepTime = Duration.ZERO
     playerController.setVolume(1F)
-    playStateManager.sleepAtEoc = false
+    sleepAtEoc = false
   }
 }
